Implement save of ICO format files to both ficons and `.ico` files
- Add a new ficonfactory::make_ficon method enabling direct setting of
formats and sizes where these can't be determined from an Image object,
especially where the Blob contains data from an ICO format file.
- Use 'std::pair's and typedefs for them to keep the argument interface
of the new make_ficon method reasonably simple.
- Rename IconsRetriever::make_ficons to push_ico_img_ficons and add more
functionality from pullICO to simplify pullICO and the
push_ico_img_ficons interface.
- Add IconsRetriever::push_ico_ficon to save the ICO file data to a
ficon.
- Add IconsRetriever::saveLargerSize to find the largest size of image
in the ICO file, for use as sizes in the ICO ficon data.
- Change the curl.pull in IconsRetriever::pullICO to use a ostringstream
rather than a file handle so the stream can be saved to a ficon's Blob.
- Implement `.ico` file saving in main.cpp.