Implement retrieval of M$ tile images specified in an XML file
- Use the Xerces API to extract image location information from the XML
file.
- Add initialization of the Xerces system within the DOMParser.
- Add method to DOMParser to enable retrieving a pointer to the
DOMDocument at any time after the document has been parsed.
- Modify the parse method to perform parsing based on a std::string url
rather than a Xerces string, thereby simplifying string handling.
- Still need to ensure that a pulled url is always trimmed down to the
domain before appending an extracted image location path lacking a
domain.
- For the time being, retain a large amount of commented-out code that
was added in the struggle to get things working correctly and/or while
debugging. (The main issue turned out to be a stupid failure to close
the XML file before trying to parse it...) Most, if not all, of this
code will most be removed later in a separate cleanup commit.